requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Session==0.5.0
psycopg2-binary==2.9.9
redis==5.0.1
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0